Module hash::crc32

Module crc32 implements the 32-bit cyclic redundancy check, or CRC-32, checksum.

See https://en.wikipedia.org/wiki/Cyclic_redundancy_check for information.

Polynomials are represented in LSB-first form also known as reversed representation.

See https://en.wikipedia.org/wiki/Mathematics_of_cyclic_redundancy_checks#Reversed_representations_and_reciprocal_polynomials for information.

Structs

IEEE_TABLE

IEEE_TABLE is the table for the IEEE polynomial.

Table

Table is a 256-word table representing the polynomial for efficient processing.

Constants

CASTAGNOLI

Castagnoli’s polynomial, used in iSCSI. Has better error detection characteristics than IEEE. https://dx.doi.org/10.1109/26.231911

IEEE

IEEE is by far and away the most common CRC-32 polynomial. Used by ethernet (IEEE 802.3), v.42, fddi, gzip, zip, png, …

KOOPMAN

Koopman’s polynomial. Also has better error detection characteristics than IEEE. https://dx.doi.org/10.1109/DSN.2002.1028931

SIZE

The size of a CRC-32 checksum in bytes.

Functions

checksum

checksum returns the CRC-32 checksum of data using the polynomial represented by the Table.

checksum_ieee

checksum_ieee returns the CRC-32 checksum of data using the IEEE polynomial.

make_table

make_table returns a Table constructed from the specified polynomial. The contents of this Table must not be modified.

new

new creates a new Hash32 computing the CRC-32 checksum using the polynomial represented by the Table. Its sum method will lay the value out in big-endian byte order.

new_ieee

new_ieee creates a new Hash32 computing the CRC-32 checksum using the IEEE polynomial. Its sum method will lay the value out in big-endian byte order.

update

update returns the result of adding the bytes in p to the crc.