Module hash::crc64

Module crc64 implements the 64-bit cyclic redundancy check, or CRC-64, checksum.

See https://en.wikipedia.org/wiki/Cyclic_redundancy_check for information.

Structs

Table

Table is a 256-word table representing the polynomial for efficient processing.

Constants

ECMA

The ECMA polynomial, defined in ECMA 182.

ISO

The ISO polynomial, defined in ISO 3309 and used in HDLC.

SIZE

The size of a CRC-64 checksum in bytes.

Functions

checksum

checksum returns the CRC-64 checksum of data using the polynomial represented by the Table.

make_table

make_table returns a Table constructed from the specified polynomial. The contents of this Table must not be modified.

new

new creates a new hash::Hash64 computing the CRC-64 checksum using the polynomial represented by the Table. Its sum method will lay the value out in big-endian byte order.

update

update returns the result of adding the bytes in p to the crc.