Module hash::fnv

Module fnv implements FNV-1 and FNV-1a, non-cryptographic hash functions created by Glenn Fowler, Landon Curt Noll, and Phong Vo.

See https://en.wikipedia.org/wiki/Fowler-Noll-Vo_hash_function.

Functions

new32

new32 returns a new 32-bit FNV-1 Hash. Its sum method will lay the value out in big-endian byte order.

new32a

new32a returns a new 32-bit FNV-1a Hash. Its sum method will lay the value out in big-endian byte order.

new64

new64 returns a new 64-bit FNV-1 Hash. Its sum method will lay the value out in big-endian byte order.

new64a

new64a returns a new 64-bit FNV-1a Hash. Its sum method will lay the value out in big-endian byte order.

new128

new128 returns a new 128-bit FNV-1 Hash. Its sum method will lay the value out in big-endian byte order.

new128a

new128a returns a new 128-bit FNV-1a Hash. Its sum method will lay the value out in big-endian byte order.