Struct hash::maphash::Hash

pub struct Hash { /* private fields */ }

A Hash computes a seeded hash of a byte sequence.

The zero Hash is a valid Hash ready to use. A zero Hash chooses a random seed for itself. For control over the seed, use set_seed.

The computed hash values depend only on the initial seed and the sequence of bytes provided to the Hash object, not on the way in which the bytes are provided. For example, the three sequences

h.write(&[b'f', b'o', b'o']);
h.write_byte(b'f'); h.write_byte(b'o'); h.write_byte(b'o');
h.write_string("foo");

all have the same effect.

Hashes are intended to be collision-resistant, even for situations where an adversary controls the byte sequences being hashed.

A Hash is not safe for concurrent use by multiple goroutines, but a Seed is. If multiple goroutines must compute the same seeded hash, each can declare its own Hash and call set_seed with a common Seed.

Example

use std::io::Write;

use hash::{maphash::Hash as MapHash, Hash, Hash64};

fn main() {
    let mut h = MapHash::new();

    h.write_string("hello, ").unwrap();
    println!("{:#016x}", h.sum64());

    h.write(&[b'w', b'o', b'r', b'l', b'd']).unwrap();
    println!("{:#016x}", h.sum64());

    h.reset();

    let mut h2 = MapHash::new();
    h2.set_seed(*h.seed());

    h.write_string("same").unwrap();
    h2.write_string("same").unwrap();

    println!("{:#016x} == {:#016x}", h.sum64(), h2.sum64());
}

Implementations

impl Hash

pub fn new() -> Self

pub fn seed(&self) -> &Seed

seed returns h’s seed value.

pub fn set_seed(&mut self, s: Seed)

set_seed sets h to use seed, which must have been returned by make_seed or by another Hash’s seed method.

Two Hash objects with the same seed behave identically. Two Hash objects with different seeds will very likely behave differently. Any bytes added to h before this call will be discarded.

pub fn write_byte(&mut self, b: u8) -> Result<()>

write_byte adds b to the sequence of bytes hashed by h. It never fails.

pub fn write_string<S>(&mut self, s: S) -> Result<usize>where
S: AsRef<str>,

write_string adds the bytes of s to the sequence of bytes hashed by h. It always writes all of s and never fails.