Trait hash::Hash

pub trait Hash: Write {
    fn sum(&mut self, b: Option<Vec<u8>>) -> Vec<u8>;
    fn reset(&mut self);
    fn size(&self) -> isize;
    fn block_size(&self) -> isize;
}

Hash is the common interface implemented by all hash functions.

std::io::Write trait requirement adds more data to the running hash. It never returns an error.

Required Methods

fn sum(&mut self, b: Option<Vec<u8>>) -> Vec<u8>

sum appends the current hash to b and returns the resulting slice. It does not change the underlying hash state.

fn reset(&mut self)

reset resets the Hash to its initial state.

fn size(&self) -> isize

size returns the number of bytes Sum will return.

fn block_size(&self) -> isize

block_size returns the hash’s underlying block size. The write method must be able to accept any amount of data, but it may operate more efficiently if all writes are a multiple of the block size.